+++ src/player.cpp
@@ -83,7 +83,7 @@ void Player::logic()
            break;
 
         case WALK:
-            mFrame = (get_elapsed_time(mWalkTime) * 6) / getWalkSpeed();
+            mFrame = (int) ((get_elapsed_time(mWalkTime) * 6) / getWalkSpeed());
             if (mFrame >= 6)
                 nextStep();
             break;
